Crate clif_util

Modules

bugpoint

CLI tool to reduce Cranelift IR files crashing during compilation.

cat

The cat sub-command.

compile

CLI tool to read Cranelift IR files and compile them into native code.

disasm

interpret

CLI tool to interpret Cranelift IR files.

print_cfg

The print-cfg sub-command.

run

CLI tool to compile Cranelift IR files to native code in memory and execute them.

souper_harvest

souper_to_peepmatic

utils

Utility functions.

wasm

CLI tool to use the functions provided by the cranelift-wasm crate.

Structs

CompiledWithoutSupportOptions

(Compiled without support for this subcommand)

PassOptions

Run specified pass(es) on an input file.

TestOptions

Run Cranelift tests

Enums

Commands

Cranelift code generator utility.

Functions

handle_debug_flag

main