Module clif_util::bugpoint

CLI tool to reduce Cranelift IR files crashing during compilation.

Structs

CrashCheckContext

MergeBlocks

MoveInstToEntryBlock

Try to move instructions to entry block.

Options

Reduce size of clif file causing panic during compilation.

RemoveBlock

Try to remove a block.

RemoveInst

Try to remove instructions.

RemoveUnusedEntities

Try to remove unused entities.

ReplaceBlockParamWithConst

Try to replace the block params with constants.

ReplaceInstWithConst

Try to replace instructions with iconst or fconst.

ReplaceInstWithTrap

Try to replace instructions with trap.

Enums

CheckResult

ProgressStatus

Traits

Mutator

Functions

block_count

const_for_type

get_panic_string

inst_count

next_inst_ret_prev

reduce

resolve_aliases

run

try_resolve_aliases

Resolve aliases only if function still crashes after this.