Trait clif_util::bugpoint::Mutator

trait Mutator {
    // Required methods
    fn name(&self) -> &'static str;
    fn mutation_count(&self, func: &Function) -> usize;
    fn mutate(
        &mut self,
        func: Function
    ) -> Option<(Function, String, ProgressStatus)>;

    // Provided method
    fn did_crash(&mut self) { ... }
}

Required Methods

fn name(&self) -> &'static str

fn mutation_count(&self, func: &Function) -> usize

fn mutate( &mut self, func: Function ) -> Option<(Function, String, ProgressStatus)>

Provided Methods

fn did_crash(&mut self)

Gets called when the returned mutated function kept on causing the crash. This can be used to update position of the next item to look at. Does nothing by default.