Crate clif_util

Modules

bugpoint ğŸ”’

CLI tool to reduce Cranelift IR files crashing during compilation.

cat ğŸ”’

The cat sub-command.

compile ğŸ”’

CLI tool to read Cranelift IR files and compile them into native code.

disasm ğŸ”’

interpret ğŸ”’

CLI tool to interpret Cranelift IR files.

print_cfg ğŸ”’

The print-cfg sub-command.

run ğŸ”’

CLI tool to compile Cranelift IR files to native code in memory and execute them.

souper_harvest ğŸ”’

utils ğŸ”’

Utility functions.

wasm ğŸ”’

CLI tool to use the functions provided by the cranelift-wasm crate.

Structs

CompiledWithoutSupportOptions ğŸ”’

(Compiled without support for this subcommand)

PassOptions ğŸ”’

Run specified pass(es) on an input file.

TestOptions ğŸ”’

Run Cranelift tests

Enums

Commands ğŸ”’

Cranelift code generator utility.

Functions

main ğŸ”’