Crate clif_util

Modules

bugpoint
CLI tool to reduce Cranelift IR files crashing during compilation.

cat
The cat sub-command.

compile
CLI tool to read Cranelift IR files and compile them into native code.

disasm

interpret
CLI tool to interpret Cranelift IR files.

print_cfg
The print-cfg sub-command.

run
CLI tool to compile Cranelift IR files to native code in memory and execute them.

utils
Utility functions.

wasm
CLI tool to use the functions provided by the cranelift-wasm crate.

Functions

add_check_translation_flag

add_debug_flag

add_disasm_flag

add_input_file_arg

add_just_decode_flag

add_pass_arg

add_print_flag

add_set_flag

add_single_input_file_arg

add_size_flag

add_target_flag

add_time_flag

add_verbose_flag

add_wasm_or_compile

get_vec
Returns a vector of clap value options and changes these options into a vector of strings

handle_debug_flag

main

Type Definitions

CommandResult
A command either succeeds or fails with an error message.