Module clif_util::bugpoint

CLI tool to reduce Cranelift IR files crashing during compilation.

Structs

CrashCheckContext ğŸ”’

MergeBlocks ğŸ”’

MoveInstToEntryBlock ğŸ”’

Try to move instructions to entry block.

Options

Reduce size of clif file causing panic during compilation.

RemoveBlock ğŸ”’

Try to remove a block.

RemoveInst ğŸ”’

Try to remove instructions.

RemoveUnusedEntities ğŸ”’

Try to remove unused entities.

ReplaceBlockParamWithConst ğŸ”’

Try to replace the block params with constants.

ReplaceInstWithConst ğŸ”’

Try to replace instructions with iconst or fconst.

ReplaceInstWithTrap ğŸ”’

Try to replace instructions with trap.

Enums

CheckResult ğŸ”’

ProgressStatus ğŸ”’

Traits

Mutator ğŸ”’

Functions

block_count ğŸ”’

get_panic_string ğŸ”’

inst_count ğŸ”’

next_inst_ret_prev ğŸ”’

reduce ğŸ”’

replace_with_const ğŸ”’

run

try_resolve_aliases ğŸ”’

Resolve aliases only if function still crashes after this.